+++ cmd/gov-pass-tray/main_nonwindows.go
@@ -1,0 +1,14 @@
+//go:build !windows
+
+package main
+
+import (
+	"fmt"
+	"os"
+)
+
+func main() {
+	fmt.Fprintln(os.Stderr, "gov-pass tray UI is only supported on Windows.")
+	os.Exit(1)
+}
+